server/requirements.txt
click==6.7
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
PyJWT==1.6.0
six==1.11.0
SQLAlchemy==1.2.4
Werkzeug==0.14.1
